dog-viewer - move dog-viewer.tsx into dog-viewer directory
